:hammer: Add a dedicated type for holding location data
