+++ src/dhv/messages/location.py
@@ -9,6 +9,10 @@ from dataclasses import dataclass
 from textual.message import Message
 from textual.widget import Widget
 
+##############################################################################
+# Local imports.
+from ..types import Location
+
 
 ##############################################################################
 @dataclass
@@ -17,36 +21,13 @@ class LocationChanged(Message):
 
     changer: Widget
     """The widget responsible for changing the location."""
-    start_line: int | None
-    """The starting line."""
-    start_column: int | None = None
-    """The starting column offset within the starting line."""
-    end_line: int | None = None
-    """The ending line."""
-    end_column: int | None = None
-    """The ending column within the line."""
+    location: Location
+    """The new location."""
 
     @property
     def control(self) -> Widget:
         """An alias for `changer`."""
         return self.changer
-
-    @property
-    def line_number(self) -> int:
-        """Alias for `start_line`."""
-        return self.start_line or 0
-
-    @property
-    def line_number_only(self) -> bool:
-        """Do we only have a line number?"""
-        return self.start_line is not None and all(
-            location is None
-            for location in (
-                self.start_column,
-                self.end_line,
-                self.end_column,
-            )
-        )
 
 
 ### location.py ends here
