+++ src/dhv/widgets/ast.py
@@ -20,6 +20,7 @@ from textual.widgets.tree import TreeNode
 ##############################################################################
 # Local imports.
 from ..messages import LocationChanged
+from ..types import Location
 
 ##############################################################################
 ASTNode = TreeNode[Any]
@@ -73,6 +74,35 @@ class AbstractSyntaxTree(Tree[Any]):
             `True` if the value should be added, `False` if not.
         """
         return bool(value) if isinstance(value, (list, tuple)) else True
+
+    @classmethod
+    def _location_of(cls, node: ASTNode) -> Location | None:
+        """Get the location of a node in the AST.
+
+        Args:
+            node: The node in the tree to get the location of.
+
+        Returns:
+            The location, or `None` if nothing could be worked out.
+        """
+        if all(
+            hasattr(node.data, location_property)
+            for location_property in (
+                "lineno",
+                "col_offset",
+                "end_lineno",
+                "end_col_offset",
+            )
+        ):
+            return Location(
+                getattr(node.data, "lineno", None),
+                getattr(node.data, "col_offset", None),
+                getattr(node.data, "end_lineno", None),
+                getattr(node.data, "end_col_offset", None),
+            )
+        elif node.parent is not None:
+            return cls._location_of(node.parent)
+        return None
 
     @singledispatchmethod
     def _base_node(self, item: Any, to_node: ASTNode) -> ASTNode:
@@ -166,37 +196,6 @@ class AbstractSyntaxTree(Tree[Any]):
         self.clear()._add(ast, self.root).root.expand_all()
         self.select_node(self.root)
 
-    @classmethod
-    def _location_of(
-        cls, node: ASTNode
-    ) -> tuple[int | None, int | None, int | None, int | None] | None:
-        """Get the location of a node in the AST.
-
-        Args:
-            node: The node in the tree to get the location of.
-
-        Returns:
-            A tuple of the location data, or `None` if nothing could be worked out.
-        """
-        if all(
-            hasattr(node.data, location_property)
-            for location_property in (
-                "lineno",
-                "col_offset",
-                "end_lineno",
-                "end_col_offset",
-            )
-        ):
-            return (
-                getattr(node.data, "lineno", None),
-                getattr(node.data, "col_offset", None),
-                getattr(node.data, "end_lineno", None),
-                getattr(node.data, "end_col_offset", None),
-            )
-        elif node.parent is not None:
-            return cls._location_of(node.parent)
-        return None
-
     @on(Tree.NodeHighlighted)
     def _ast_node_highlighted(self, message: Tree.NodeHighlighted[ASTNode]) -> None:
         """Handle a node being highlighted in the AST.
@@ -206,7 +205,7 @@ class AbstractSyntaxTree(Tree[Any]):
         """
         if location := self._location_of(message.node):
             message.stop()
-            self.post_message(LocationChanged(self, *location))
+            self.post_message(LocationChanged(self, location))
 
 
 ### ast.py ends here
